+++ src/cpp/entities/dragger.cpp
@@ -17,7 +17,8 @@ Dragger::Dragger() {
 
 std::optional<Vector2> Dragger::update() {
   if (drag_init_pos.has_value() &&
-      Input::isMouseButtonReleased(MOUSE_LEFT_BUTTON)) {
+      (Input::isMouseButtonReleased(MOUSE_LEFT_BUTTON) ||
+       Input::isKeyReleased(KEY_SPACE))) {
     float x_diff = win::getMouseX() - drag_init_pos->x;
     float y_diff = win::getMouseY() - drag_init_pos->y;
 
@@ -33,7 +34,8 @@ std::optional<Vector2> Dragger::update() {
 
     return drag;
 
-  } else if (Input::isMouseButtonPressed(MOUSE_LEFT_BUTTON)) {
+  } else if (Input::isMouseButtonPressed(MOUSE_LEFT_BUTTON) ||
+             Input::isKeyPressed(KEY_SPACE)) {
     drag_init_pos = win::getMousePos();
   }
 
@@ -77,8 +79,8 @@ void Dragger::drawArrow(const Squircle &squircle) {
   }
 }
 
-void Dragger::drawCircle() {
+void Dragger::drawCircle(Color color) {
   if (drag_init_pos.has_value()) {
-    win::drawCircle(*drag_init_pos, 10, GREEN);
+    win::drawCircle(*drag_init_pos, 10, color);
   }
 }