+++ src/cpp/entities/hud.cpp
@@ -1,4 +1,5 @@
 #include "entities/hud.hpp"
+#include "game_state.hpp"
 #include "raylib.h"
 #include <string>
 
@@ -6,14 +7,16 @@ HeightDisplay::HeightDisplay() {
   hud_font = ResourceManager::getFont("comfortaa.ttf");
 }
 
-void HeightDisplay::draw(float height, float max_height, int jumps) {
-  DrawTextEx(*hud_font, std::to_string(int(height)).c_str(), {20, 20}, 46, 0,
-             WHITE);
-  DrawTextEx(*hud_font, std::to_string(int(max_height)).c_str(), {20, 60}, 24,
-             0, WHITE);
+void HeightDisplay::draw(std::shared_ptr<GameState> state) {
+  DrawTextEx(*hud_font, std::to_string(int(state->height)).c_str(), {20, 20},
+             46, 0, WHITE);
+  DrawTextEx(*hud_font, std::to_string(int(state->max_height)).c_str(),
+             {20, 60}, 24, 0, WHITE);
 
-  std::string jumpText = "Jumps: " + std::to_string(jumps);
-  Vector2 size = MeasureTextEx(*hud_font, jumpText.c_str(), 24, 0);
+  std::string jumpText = "Jumps: " + std::to_string(state->jumps_made);
+  DrawTextEx(*hud_font, jumpText.c_str(), {20, 140}, 24, 0, WHITE);
 
-  DrawTextEx(*hud_font, jumpText.c_str(), {20, 100}, 24, 0, WHITE);
+  std::string doubleJumpText =
+      "Doubles: " + std::to_string(state->double_jumps);
+  DrawTextEx(*hud_font, doubleJumpText.c_str(), {20, 100}, 24, 0, WHITE);
 }