+++ src/cpp/entities/objects/objects.cpp
@@ -1,6 +1,7 @@
 #include "entities/objects/objects.hpp"
 
 #include "entities/objects/basic_platform.hpp"
+#include "entities/objects/double_jump_item.hpp"
 #include "entities/objects/ghost_platform.hpp"
 #include "entities/objects/icy_platform.hpp"
 #include "entities/objects/spiky_platform.hpp"
@@ -19,6 +20,8 @@ ObjectP createObject(std::string id) {
     return std::make_unique<GhostPlatform>();
   } else if (id == TextObject::id) {
     return std::make_unique<TextObject>();
+  } else if (id == DoubleJumpItem::id) {
+    return std::make_unique<DoubleJumpItem>();
   }
 
   return nullptr;
@@ -40,6 +43,9 @@ ObjectP copyObject(Object *object) {
         *dynamic_cast<GhostPlatform *>(object));
   } else if (id == TextObject::id) {
     return std::make_unique<TextObject>(*dynamic_cast<TextObject *>(object));
+  } else if (id == DoubleJumpItem::id) {
+    return std::make_unique<DoubleJumpItem>(
+        *dynamic_cast<DoubleJumpItem *>(object));
   }
 
   return nullptr;